Implemented JWT authentication
Added a filter which checks for the Authorization header in the request
and if it exists, extracts the token from it, validates it and extracts
the user information from it, attaches it to the security context as the
authenticated principal.
The token is retrieved through the users/login endpoint upon successful
authentication with a username and password. It is returned in the
response body along with other meta data (expiration time and type).
The JwtUtil class is used for managing tokens. It uses an application
secret key and expiration time configurations from the
application.properties file.
Currently the JWT holds as user information only the user id from the
users table - as the token subject.