Create auth service, JWT interceptor, login logic
Created an authentication service, which stores information about the
currently logged in user and exposes an observable, to which other parts
of the application can subscribe to in order to respond to changes in
the currently logged in user.
It uses a token service, which handles the JWT storing and parsing. The
JWT is stored in local storage, so the user can stay logged in amidst
changing browser tabs and windows, as long as the token is not expired.
The token service exposes the JWT body decoded data for use in the
authentication service.
The JWT interceptor intercepts outgoing requests and adds an
Authorization header, containing the JWT, as long as it exists and is
valid, and the request is targeting the server app.
The login component uses reactive forms and clarity design.