Implement charity creation logic
Create the charity creation form.
Include a depedency on ngx-currency in order to display the currency input
field of the form better.
On uploading an image, the content is extracted in the form of binary
array in order to be sent to the backend.
To be included: custom validator to check if at least one of amount
required and volunteers required is not zero.