Implement delete charity action
Opens a modal confirmation dialog.
The button is not added to the DOM unless the current user is the charity owner.
Fixed a bug related to getting the current user from the token.
Renamed the created_at charity property to createdAt to follow conventions.
Fixed a bug related to validating the number inputs when creating
charity.