Implement volunteer endpoint
Created a volunteering repository, service and controller.
Used a transaction to handle multiple writes to the database in a single operation -
the write to the volunteerings table and the update of the volunteers count
in the charity.
Returns 403 FORBIDDEN if charity doesn't need more volunteers or if
the current user has already volunteered.
Extracted the method getting currently authenticated user to the user service.