Implement user donation on the backend
A user can donate only if the amount is not more than the remaining to collect.
Otherwise it returns 400 BAD REQUEST.
Refactored the volunteering controller and service to also return 400 on
non valid volunteering requests for consistency (instead of 403 FORBIDDEN).
Added more validation to some DTOs.
Since the donation operation requires multiple read/write accesses to the
database, the whole function is executed as a single transaction with the
@Transactional annotation