Add create charity button and search box
Whenever the search box is left untouched for 200 milliseconds,
the search term is sent to the backend via query parameters
to retrieve the filtered list of charities.
The switchMap operator is used to only process the results from the latest search
instead of all old ones.
The create charity button is shown only to authenticated users.